Functions

+++ src/interpret/linux_x86_64/syscalls.rs
@@ -1,10 +1,10 @@
-
-pub fn sys_write(sys_n: usize, fd: usize, buff: usize, count: usize, mem: &Vec<u8> ) -> usize {
+#[allow(clippy::cast_possible_truncation)]
+pub fn sys_write(sys_n: usize, fd: usize, buff: usize, count: usize, mem: &Vec<u64> ) -> usize {
     let mem = (*mem).clone();
     // println!("{:?}", &mem[buff..(buff + count)]);
     // return 0 ;
     let s = &mem[buff..(buff + count)].iter().map(|i| {
-        char::from_u32(u32::from(*i)).unwrap_or('_').to_string()
+        char::from_u32(u32::from(*i as u8)).unwrap_or('_').to_string()
     }).collect::<String>();
     
     match fd {
